move ui library to a folder
